moved commonHandlers to it's own file
