+++ index.js
@@ -3,14 +3,7 @@ const router = express.Router();
 const requestParser = require('./api/requestParser');
 const memeBuilder = require('./api/memebuilder');
 
-const commonHandlers = {
-  e405: function(_req, res, _next) {
-    res.status(405).send('This Api only accepts GET requests.');
-  },
-  e404: function (_req, res, _next) {
-    res.status(404).send('File Not Found');
-  }
-};
+const commonHandlers =  require('./commonHandlers');
 
 async function memeHandler(req, res, _next) {
   const memeName = req.params['memeName'];
